chore: consolidate relative-time formatters into shared util
formatRelativeTime in src/lib/utils.ts now accepts number | string | Date,
covering both the AI session-history use case (ISO strings) and the new
port-forward history use case (epoch ms). The old AI-local formatRelativeTime
is removed; ai/lib/utils.ts re-exports it for stable import paths.

Drops the redundant formatTimeAgoShort I added earlier in the
port-forward history work — same behavior, but with the 7d→date fallback
that the AI version already had.
